requirements/common.txt
requests==2.31.0
pandas==2.0.1
python-dotenv==1.0.0
azure-functions==1.17.0
azure-functions-durable==1.2.8
azure-cosmos == 4.3.1
jsonschema == 4.17.3
stravalib == 1.5